Rename restApiSearchClient.ts to restSearchClient.ts
