+++ src/components/typeahead-search/http/restSearchClient.ts
